+++ site/dat/docs/Changelog.md
@@ -4,6 +4,7 @@
  - fix shellexec env variables
  - add `assume-success` flag for JMeter assertions
  - flexibly handle '$1$' templates for JMeter regex extractor
+ - fix null body parameter for JMeter
 
 ## 0.5.0
  - allow using Xvfb for Selenium executor